Make environment vars available in scripts opt-in
This  script outputs the environment in which the other scripts should run. Its purpose is to ensure that build scripts only have access to the environment variables they need to do their work. This forces us to opt-in to providing an environment variable to a portion of the build process. This should prevent secrets from leaking into parts of the build process where they should not be accessible.

Example:

. /app/environment.sh clone
GITHUB_TOKEN="123abc" SOURCE_OWNER="18F" SOURCE
REPO="modern-team-template" ...

This is intended to be used with `env` to restrict the environment of the other scrips.